refactor(platform-roms): split PlatformROMsListView into separate files
- SyncSaveSheet.swift: SyncSaveSheet + PendingUpload
- ShareSwipeButton.swift: ShareSwipeButton + ShareSheetItem
- ROMCardRow.swift: ROMCardRow, SaveBadge, CardActionButton, CardActionLabel
- PlatformROMsListView.swift: main view only
